+++ src/main/java/org/highfive/backend/auth/service/AuthService.java
@@ -15,7 +15,7 @@ import org.highfive.backend.global.dto.Response;
 import org.highfive.backend.global.exception.BusinessException;
 import org.highfive.backend.user.entity.Role;
 import org.highfive.backend.user.entity.User;
-import org.highfive.backend.user.mapper.UserMapper;
+import org.highfive.backend.user.dto.mapper.UserMapper;
 import org.highfive.backend.user.repository.UserRepository;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.stereotype.Service;
